requirements.txt
streamlit 
pandas
numpy
matplotlib
plotly
tweepy
python-dotenv
altair
ccxt